// src/feature-module/peoplesManagement/type.ts
export interface CustomerDataType {
    [key: string]: any;
    name: string;
    mobile: string;
    email: string;
    address: string;
    pan: string;
    gst: string;
  }
export interface StoreDataType {
    [key: string]: any;
    stateName: string;
    storeCode: string;
    description: string;
   
  }